+++ app/assets/scripts/minireel/app.js
@@ -323,6 +323,19 @@ function( angular , c6uilib , c6State  , services          , tracker          ,
             };
         }])
 
+        .filter('embedid', ['VideoService',
+        function           ( VideoService ) {
+            var embedIdFromVideoId = VideoService.embedIdFromVideoId;
+
+            function reverse(array) {
+                return Array.prototype.reverse.call(array);
+            }
+
+            return function() {
+                return embedIdFromVideoId.apply(null, reverse(arguments));
+            };
+        }])
+
         .directive('paginatorControls', [function() {
             return {
                 scope: {
@@ -572,6 +585,7 @@ function( angular , c6uilib , c6State  , services          , tracker          ,
             var minireelState = c6State.get('MiniReel'),
                 categories = minireelState.cModel.data.modes,
                 c6EmbedSrc = minireelState.cModel.data.c6EmbedSrc,
+                soloPlayerUrl = minireelState.cModel.data.soloPlayerUrl,
                 orgSettings = SettingsService.getReadOnly('MR::org');
 
             var allFormats = {
@@ -582,6 +596,10 @@ function( angular , c6uilib , c6State  , services          , tracker          ,
                 shortcode: {
                     name: 'Wordpress Shortcode',
                     value: 'shortcode'
+                },
+                iframe: {
+                    name: 'IFrame',
+                    value: 'iframe'
                 }
             };
 
@@ -662,6 +680,16 @@ function( angular , c6uilib , c6State  , services          , tracker          ,
                         return formatEmbed(
                             '<script src="' + c6EmbedSrc + '" |data-{attr}="{value}"|></script>',
                             data
+                        );
+                    case 'iframe':
+                        return formatEmbed(
+                            '<iframe |{attr}="{value}"|></iframe>',
+                            {
+                                src: soloPlayerUrl + '?id=' + minireel.id,
+                                frameborder: '0',
+                                width: explicitDimensions ? this.size.width : '100%',
+                                height: explicitDimensions ? this.size.height : '100%'
+                            }
                         );
                     }
                 }