+++ test/spec/minireel/video.service.ut.js
@@ -53,6 +53,12 @@
                             expect(fromData('yahoo', 'vote-marijuana-034413930')).toBe('https://screen.yahoo.com/vote-marijuana-034413930.html');
                             expect(fromData('yahoo', 'gopro-shows-whisky-bottle-being-085523254')).toBe('https://screen.yahoo.com/gopro-shows-whisky-bottle-being-085523254.html');
                         });
+
+                        it('should create a rumble url', function() {
+                            ['v2zfax-my-cat-dog-meets-mac-n-cheese-for-the-first-time', 'v2z8ro-willie-perfoming-at-school-talent-show', 'v2zfat-white-german-shepherd-and-baby-goat'].forEach(function(videoid) {
+                                expect(fromData('rumble', videoid)).toBe('https://rumble.com/' + videoid + '.html');
+                            });
+                        });
                     });
 
                     describe('dataFromUrl(url)', function() {
@@ -95,6 +101,13 @@
                             });
                         });
 
+                        it('should parse a rumble url', function() {
+                            expect(fromUrl('https://rumble.com/v2zfax-my-cat-dog-meets-mac-n-cheese-for-the-first-time.html')).toEqual({
+                                service: 'rumble',
+                                id: 'v2zfax-my-cat-dog-meets-mac-n-cheese-for-the-first-time'
+                            });
+                        });
+
                         it('should return null if the url is not valid', function() {
                             expect(fromUrl('apple.com')).toBeNull();
                             expect(fromUrl('84fh439#')).toBeNull();
@@ -102,6 +115,32 @@
                             expect(fromUrl('http://www.vimeo.com/')).toBeNull();
                             expect(fromUrl('http://www.dailymotion.com/')).toBeNull();
                             expect(fromUrl('http://www.youtube.c')).toBeNull();
+                        });
+                    });
+
+                    describe('embedIdFromVideoId(service, videoid)', function() {
+                        var ids = ['8439htf4', '8394tr8u394r', '2984ru43', '4892ur43', 'r892ur4'];
+
+                        ['yahoo', 'vimeo', 'dailymotion', 'aol', 'yahoo'].forEach(function(service, index) {
+                            describe('for ' + service, function() {
+                                var result;
+
+                                beforeEach(function() {
+                                    result = VideoService.embedIdFromVideoId(service, ids[index]);
+                                });
+
+                                it('should return the videoid', function() {
+                                    expect(result).toBe(ids[index]);
+                                });
+                            });
+                        });
+
+                        describe('for rumble', function() {
+                            it('should be the properly formed embed id', function() {
+                                expect(VideoService.embedIdFromVideoId('rumble', 'v2zfat-white-german-shepherd-and-baby-goat')).toBe('8.2zfat');
+                                expect(VideoService.embedIdFromVideoId('rumble', 'v2zfbw-police-nutbush-g20-brisbane')).toBe('8.2zfbw');
+                                expect(VideoService.embedIdFromVideoId('rumble', 'v2z8ro-willie-perfoming-at-school-talent-show')).toBe('8.2z8ro');
+                            });
                         });
                     });