+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # MiniReelinator
 
+## v1.6.0 (December 8, 2014)
+* *[v1.6.0.rc1]*
+  * Changed tab order for new card wizard
+  * [FEATURE]: Added ability to render embed code for MiniReel.tv
+    (iframe embed code)
+  * [FEATURE]: Added ability to add Rumble.com videos to MiniReels
+* *[/v1.6.0.rc1]*
+
 ## v1.5.3 (December 4, 2014)
 * *[v1.5.3.rc1]*
   * When a MR is copied, the cards in that MR will be given new IDs